--- a/optimum/habana/transformers/models/__init__.py
+++ b/optimum/habana/transformers/models/__init__.py
@@ -190,8 +190,6 @@ from .starcoder2 import (
     GaudiStarcoder2DecoderLayer,
     GaudiStarcoder2ForCausalLM,
     GaudiStarcoder2Model,
-    #gaudi_starcoder2_attention_forward,
-    #gaudi_starcoder2_model_forward,
 )
 from .swin import gaudi_swin_get_attn_mask
 from .t5 import (