--- a/optimum/habana/transformers/models/starcoder2/__init__.py
+++ b/optimum/habana/transformers/models/starcoder2/__init__.py
@@ -3,6 +3,4 @@ from .modeling_starcoder2 import (
     GaudiStarcoder2ForCausalLM,
     GaudiStarcoder2Attention,
     GaudiStarcoder2Model,
-    #gaudi_starcoder2_attention_forward,
-    #gaudi_starcoder2_model_forward,
 )