# What does this PR do?

Adds KVCache implementation to the Starcoder2 model 
Adds Gaudi Flash Attention (SDPA) to StarCoder2 model.

Implementation borrowed from Qwen2 PRs :
[#1087](https://github.com/huggingface/optimum-habana/pull/1087)
[#1033](https://github.com/huggingface/optimum-habana/pull/1033)  


Validation on optimum-habana  text-generation inference : 
===========================================

> **With PR :** 
> `python run_generation.py --model_name_or_path "bigcode/starcoder2-3b"  --bf16 --batch_size 100 --use_hpu_graphs   --do_sample   --prompt "def print_hello_world():"   --max_new_tokens 550  --max_input_tokens 30    --use_kv_cache  --reuse_cache    --use_flash_attention  --flash_attention_recompute`
> 
> ```
> 07/22/2024 16:32:50 - INFO - __main__ - device: hpu, n_hpu: 0, bf16: True
> 07/22/2024 16:32:50 - INFO - __main__ - Model initialization took 8.926s                                                                                                                                                                  07/22/2024 16:32:50 - INFO - __main__ - Graph compilation...                                                                                                                                                                              Warming up
> 07/22/2024 16:32:56 - INFO - __main__ - Time to first token = 1038.1485800025985ms                                                                                                                                                        Warming up
> 07/22/2024 16:33:01 - INFO - __main__ - Time to first token = 107.46545001165941ms
> Warming up
> 07/22/2024 16:33:05 - INFO - __main__ - Time to first token = 64.5028020371683ms
> 07/22/2024 16:33:05 - INFO - __main__ - Running generate...
> 07/22/2024 16:33:10 - INFO - __main__ - Time to first token = 63.07519105030224ms
> 07/22/2024 16:33:14 - INFO - __main__ - Time to first token = 63.960200001019984ms
> 07/22/2024 16:33:18 - INFO - __main__ - Time to first token = 65.7379609765485ms
> 07/22/2024 16:33:23 - INFO - __main__ - Time to first token = 65.4228410567157ms
> 07/22/2024 16:33:27 - INFO - __main__ - Time to first token = 64.82256192248315ms
> 
> input 1: ('def print_hello_world():',)
> output 1: ('def print_hello_world():\n    """Print Hello World\n    """\n    print("hello world!")\nfrom setuptools import setup, find_packages\n\nsetup(\n    name="HelloWorld",\n    version="0.1.0",\n    python_requires=">3.7",\n
> description="This is Hello World.",\n    packages=find_packages()\n)\nfrom setuptools import find_packages, setup, Extension\n\nextensions = [\n    Extension(\n        name = "say", \n        sources = ["scripts/say/say.c"],\n    ),\n]\n\nsetup(\n    name = "say",\n    version = "1.0",\n    description = "say in python",\n    author = "",\n    license = "MIT",\n    packages = find_packages(),\n    ext_package = "saylib",\n    ext_modules = extensions,\n)\npackage
> main\n\nimport "fmt"\n\nfunc main() {\n    println("hello world!")\n}\nfrom setuptools import setup, find_packages\n\nsetup(\n    name="saylib",\n    version="2.0.0",\n    description="Say Lib",\n    license="MIT",\n    packages=find_packages(),\n    package_data={\n       \'saylib\': [\'say.txt\']\n    },\n    include_package_data=True,\n)\nfrom setuptools import setup, Extension\n\ndef main():\n    setup(\n        name="hello_world",\n        version="1.0",\n        ext_modules=[\n            Extension("say",\n                [\'say.c\']\n            ),\n        ],\n    )\n\nmain()\n# python - setuptools________\n\n[![python-hello-world](https://github.com/shin1ogari/python-hello-world/workflows/python-hello-world/badge.svg)](PYTHON-HELLO-WORLD)\n\nhttps://shin1ogari.hatenablog.com/entry/2021/06/17/154143\n\n## How to Install\n\n```\ngit clone https://github.com/shin1ogari/python-hello-world.git\ncd python-hello-world\npython3 setup.py sdist\npython3 -m pip install dist/HelloWorld-0.1.0.tar.gz\n```\n/**\n * @author \n */\n\n/*This function will give the list of tasks in 2 lists, one completed and one not completed\n * @todo : Sort the list on basis of due date \n */\n\n\nconst tasks = [\n    {\n        taskName: "',)
> 
> Stats:
> ---------------------------------------------------------------------------------------------------------------
> Throughput (including tokenization) = 12568.517464337727 tokens/second
> Number of HPU graphs                      = 19
> Memory allocated                               = 9.11 GB
> Max memory allocated                       = 10.24 GB
> Total memory available                       = 94.62 GB
> Graph compilation duration                = 14.794863908027764 seconds
> ---------------------------------------------------------------------------------------------------------------
> ```
> 
> 
> **Without PR :** 
> `python run_generation.py --model_name_or_path "bigcode/starcoder2-3b"  --bf16 --batch_size 100 --use_hpu_graphs   --do_sample   --prompt "def print_hello_world():"   --max_new_tokens 550  --max_input_tokens 30    --use_kv_cache   --use_flash_attention  --flash_attention_recompute`
> 
> ```
> 07/22/2024 16:15:16 - INFO - __main__ - device: hpu, n_hpu: 0, bf16: True                                                                                                                                                                 07/22/2024 16:15:16 - INFO - __main__ - Model initialization took 8.410s                                                                                                                                                                  07/22/2024 16:15:16 - INFO - __main__ - Graph compilation...                                                                                                                                                                              Warming up
> 07/22/2024 16:15:51 - INFO - __main__ - Time to first token = 28370.036993990652ms
> Warming up                                                                                                                                                                                                                                07/22/2024 16:15:57 - INFO - __main__ - Time to first token = 1484.4204590190202ms
> Warming up
> 07/22/2024 16:16:03 - INFO - __main__ - Time to first token = 1472.5127129931934ms                                                                                                                                                        07/22/2024 16:16:03 - INFO - __main__ - Running generate...                                                                                                                                                                               07/22/2024 16:16:09 - INFO - __main__ - Time to first token = 1471.8866730108857ms                                                                                                                                                        07/22/2024 16:16:15 - INFO - __main__ - Time to first token = 1471.6727570048533ms
> 07/22/2024 16:16:20 - INFO - __main__ - Time to first token = 1476.9293259596452ms
> 07/22/2024 16:16:26 - INFO - __main__ - Time to first token = 1473.4082049690187ms
> 07/22/2024 16:16:32 - INFO - __main__ - Time to first token = 1473.347980005201ms
> 
> input 1: ('def print_hello_world():',)                                                                                                                                                                                               
> output 1: ('def print_hello_world():\n    print("Hello, world!")\n\nprint_hello_world() # ____\n\ndef make_square_area(size):\n    return size ** 2\n\nprint(make_square_area(5))\nprint(make_square_area(10))\nprint(make_square_area(13)
> )\n\n\n\n# def _________(____):\n#     ____ ____\n#     ____ ____\n#    ...\n#     return ______\n\n# 1. ______ 10____ ____ ________ ______ ____ (________ ____)\n# 2. ______ ________ ____ (______ ____)\n# 3. 1____ 100______ ______ __ __ ____ ________ return____ ____\n\n\n#!/bin/python3\n\nimport math\nimport os\nimport random\nimport re\nimport sys\n\n\n\n#\n# Complete the \'addIntegers\' function below.\n#\n# The function is expected to return an INTEGER.\n# The function accepts following parameters:\n#  1. INTEGER A\n#  2. INTEGER B\n#\n\n# 10 20\n# 10 20   \n# 10 + 20 = 30\n# 10 + 20 = 30\n\n#  def addIntegers(A, B):\n#     for i in range(0, 2):\n#         for j in range(0, 2):\n#
>    print(f\'{A[i]} + {B[j]} = {A[i] + B[j]}\')\n\ndef addIntegers(A, B):\n    result = []\n    for i in range(0, A):\n        for j in range(0, B):\n            sum = A[i] + B[j]\n            result.append(sum)\n    return result \n\n\n\n\nif __name__ == \'__main__\':\n    fptr = open(os.environ[\'OUTPUT_PATH\'], \'w\')\n\n    first_multiple_input = input().rstrip().split()\n\n    A = int(first_multiple_input[0])\n\n    B = int(first_multiple_input[1])\n    A_a = []\n    B_a = []\n    for i in range(0, A):\n        A_a.append(i)\n    for i in range(0, B):\n        B_a.append(i)\n    res = addIntegers',)
> 
> Stats:
> -------------------------------------------------------------------------------------------------------------
> Throughput (including tokenization) = 9344.66483451947 tokens/second
> Number of HPU graphs                = 14
> Memory allocated                    = 24.6 GB
> Max memory allocated                = 34.31 GB
> Total memory available              = 94.62 GB
> Graph compilation duration          = 46.98551622900413 seconds
> -------------------------------------------------------------------------------------------------------------
> ```