@vidyasiv 2024-07-25
@abhatkal , you might want to check/update [test_text_generation_example.py for starcoder2](https://github.com/huggingface/optimum-habana/blob/main/tests/test_text_generation_example.py#L39) to showcase the improvement with this PR.