@abhatkal 2024-07-26
@ssarkar2 @abhilash1910  I observed the below [lines](https://github.com/12010486/optimum-habana/blob/aaa230deb7b00cb2adf5f86aa0eb8f967fc596d8/optimum/habana/transformers/models/starcoder2/modeling_starcoder2.py#L615-L620 ) consistently result in bad outputs.   
            
            if (
                lazy_mode
                and not self.training
                and (torch.distributed.is_initialized() is False or torch.distributed.get_world_size() == 1)
            ):
                htcore.mark_step()
                
Commenting them gives better results without causing much difference in throughput.  However the ttft increases for the 1st warmup step.  Any better way around this ? 

Example run :    

python run_generation.py --model_name_or_path bigcode/starcoder2-3b --batch_size 1 --use_hpu_graphs   --do_sample   --prompt "def print_hello_world():"  --use_kv_cache  --reuse_cache --use_flash_attention  --flash_attention_recompute --bf16

**Results after commenting the above lines:** 

```
Warming up
07/26/2024 09:12:02 - INFO - __main__ - Time to first token = 2216.9843481387943ms
Warming up
07/26/2024 09:12:02 - INFO - __main__ - Time to first token = 65.63230406027287ms
Warming up
07/26/2024 09:12:03 - INFO - __main__ - Time to first token = 9.99257992953062ms
07/26/2024 09:12:03 - INFO - __main__ - Running generate...
07/26/2024 09:12:05 - INFO - __main__ - Time to first token = 8.498925948515534ms

input 1: ('def print_hello_world():',)
output 1: ('def print_hello_world():\n    print("Hello world")\n\n\nif __name__ == "__main__":\n    print_hello_world()/Day16/ Day 16 - Sets and Modules/ Sets and Modules/set.py\nmyset = set()\nprint(myset)\nmyset = set([1,2,3])\nprint(myset)\nmyset = set((1,2,3))\nprint(myset)\nmyset = set({1',)

Stats:
--------------------------------------------------------------------------------------------------------------
Throughput (including tokenization) = 246.3352487081534 tokens/second
Number of HPU graphs                = 17
Memory allocated                    = 6.17 GB
Max memory allocated                = 6.2 GB
Total memory available              = 94.62 GB
Graph compilation duration          = 5.175242085941136 seconds
--------------------------------------------------------------------------------------------------------------
```


**Results with the above lines intact:**  

```
Warming up
07/26/2024 09:09:29 - INFO - __main__ - Time to first token = 670.1515450840816ms
Warming up
07/26/2024 09:09:29 - INFO - __main__ - Time to first token = 53.274671896360815ms
Warming up
07/26/2024 09:09:30 - INFO - __main__ - Time to first token = 8.514783112332225ms
07/26/2024 09:09:30 - INFO - __main__ - Running generate...
07/26/2024 09:09:30 - INFO - __main__ - Time to first token = 7.96029600314796ms

Input/outputs:
input 1: ('def print_hello_world():',)
output 1: ('def print_hello_world():\n   \n.8\n       )- = `-9_ 7-)\'\',567_s. `9 =a0: "/(7_6_s_ 1 [ the\n',)

Stats:
--------------------------------------------------------------------------------------------------------------
Throughput (including tokenization) = 251.1309268210929 tokens/second
Number of HPU graphs                = 19
Memory allocated                    = 6.17 GB
Max memory allocated                = 6.18 GB
Total memory available              = 94.62 GB
Graph compilation duration          = 2.4475005640415475 seconds
-------------------------------------------------------------------------------------------------------------- 
```